Relocation of drawingTab into fileSection
