+++ cypress.config.ts
@@ -8,6 +8,7 @@ dotenv.config();
 export default defineConfig({
   projectId: "wf7d2m",
   defaultCommandTimeout: 10000,
+
   e2e: {
     setupNodeEvents(on, config) {
       // implement node event listeners here
@@ -39,8 +40,16 @@ export default defineConfig({
     numTestsKeptInMemory: 15,
     experimentalMemoryManagement: true,
   },
+
   env: {
     API_URL: process.env.REACT_CARE_API_URL,
     ENABLE_HCX: process.env.REACT_ENABLE_HCX ?? false,
+  },
+
+  component: {
+    devServer: {
+      framework: "react",
+      bundler: "vite",
+    },
   },
 });
