+++ package.json
@@ -88,7 +88,6 @@
     "clsx": "^2.1.1",
     "cmdk": "^1.1.1",
     "cross-env": "^7.0.3",
-    "cypress": "^14.2.1",
     "date-fns": "^4.0.0",
     "dayjs": "^1.11.13",
     "framer-motion": "^12.6.3",
@@ -146,6 +145,7 @@
     "@types/uuid": "^10.0.0",
     "@typescript-eslint/eslint-plugin": "^8.21.0",
     "@typescript-eslint/parser": "^8.21.0",
+    "cypress": "^14.3.2",
     "cypress-localstorage-commands": "^2.2.6",
     "cypress-split": "^1.24.5",
     "dompurify": "^3.2.5",
