+++ public/locale/en.json
@@ -2551,6 +2551,7 @@
   "search_encounters": "Search Encounters",
   "search_existing_consent": "Search existing consent",
   "search_facilities": "Search for available facilities to book an appointment <strong>(for new patients)</strong>",
+  "search_files": "Search files",
   "search_for_a_user": "Search for a user",
   "search_for_allergies_to_add": "Search for allergies to add",
   "search_for_diagnoses_to_add": "Search for diagnoses to add",
