+++ src/Routers/routes/PatientRoutes.tsx
@@ -3,7 +3,6 @@ import { Suspense, lazy } from "react";
 
 import Loading from "@/components/Common/Loading";
 import { patientTabs } from "@/components/Patient/PatientDetailsTab";
-import { PatientDrawingTab } from "@/components/Patient/PatientDetailsTab/PatientDrawingsTab";
 import { PatientHome } from "@/components/Patient/PatientHome";
 import PatientIndex from "@/components/Patient/PatientIndex";
 import PatientRegistration from "@/components/Patient/PatientRegistration";
@@ -84,9 +83,6 @@ const PatientRoutes: AppRoutes = {
         drawingId={drawingId}
       />
     </Suspense>
-  ),
-  "/patient/:patientId/drawings": ({ patientId }) => (
-    <PatientDrawingTab patientId={patientId} />
   ),
 
   "/patient/:patientId/drawings/new": ({ patientId }) => {
