+++ src/components/Common/Drawings/ExcalidrawEditor.tsx
@@ -7,7 +7,6 @@ import {
   useQuery,
   useQueryClient,
 } from "@tanstack/react-query";
-import { navigate } from "raviger";
 import { useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { toast } from "sonner";
@@ -30,6 +29,8 @@ import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 
 import Loading from "@/components/Common/Loading";
+
+import useAppHistory from "@/hooks/useAppHistory";
 
 import mutate from "@/Utils/request/mutate";
 import query from "@/Utils/request/query";
@@ -54,6 +55,7 @@ export default function ExcalidrawEditor({
   const [name, setName] = useState("");
   const [isDirty, setIsDirty] = useState(false);
   const [isAlertOpen, setIsAlertOpen] = useState(false);
+  const { goBack } = useAppHistory();
 
   const { mutate: saveDrawing } = useMutation({
     mutationFn: mutate(metaArtifactApi.upsert),
@@ -61,7 +63,7 @@ export default function ExcalidrawEditor({
       queryClient.invalidateQueries({
         queryKey: ["drawing", drawingId, associatingId],
       });
-      navigate("../drawings");
+      goBack();
     },
   });
 
@@ -113,7 +115,7 @@ export default function ExcalidrawEditor({
     if (isDirty) {
       setIsAlertOpen(true);
     } else {
-      navigate("../drawings");
+      goBack();
     }
   };
 
@@ -142,7 +144,7 @@ export default function ExcalidrawEditor({
               className="bg-red-500 text-gray-50 shadow-xs hover:bg-red-500/90"
               onClick={() => {
                 setIsAlertOpen(false);
-                navigate("../drawings");
+                goBack();
               }}
             >
               {t("discard_changes")}
