+++ src/components/Patient/PatientDetailsTab/index.tsx
@@ -7,7 +7,6 @@ import { HasPermissionFn, getPermissions } from "@/common/Permissions";
 import { Patient } from "@/types/emr/newPatient";
 
 import { Appointments } from "./Appointments";
-import { PatientDrawingTab } from "./PatientDrawingsTab";
 import { PatientFilesTab } from "./PatientFiles";
 import { PatientUsers } from "./PatientUsers";
 import { ResourceRequests } from "./ResourceRequests";
@@ -61,10 +60,6 @@ export const BASE_PATIENT_TABS: Tab[] = [
   {
     route: "files",
     component: PatientFilesTab,
-  },
-  {
-    route: "drawings",
-    component: PatientDrawingTab,
   },
 ];
