--- src/pages/Encounters/tabs/EncounterDrawingsTab.tsx
@@ -1,13 +1,0 @@
-import { DrawingTab } from "@/components/Common/Drawings/DrawingTab";
-
-import { EncounterTabProps } from "@/pages/Encounters/EncounterShow";
-
-export const EncounterDrawingsTab = (props: EncounterTabProps) => {
-  return (
-    <DrawingTab
-      type="encounter"
-      patient={props.patient}
-      encounter={props.encounter}
-    />
-  );
-};
